requirements.txt
vocode
langchain
twilio
fastapi
python_dotenv
google-search-results